refactor: merge ConnectionStringParser into ConnectionConfiguration
