+++ src/NServiceBus.RabbitMQ/Configuration/ConnectionConfiguration.cs
@@ -2,12 +2,20 @@
 {
     using System;
     using System.Collections.Generic;
+    using System.ComponentModel;
+    using System.Data.Common;
     using System.Diagnostics;
     using System.IO;
+    using System.Linq;
+    using System.Reflection;
+    using System.Text;
+    using Logging;
     using Support;
 
     class ConnectionConfiguration
     {
+        static readonly ILog Logger = LogManager.GetLogger(typeof(ConnectionConfiguration));
+
         public string Host { get; set; }
 
         public int Port { get; set; }
@@ -30,20 +38,105 @@
 
         public Dictionary<string, object> ClientProperties { get; } = new Dictionary<string, object>();
 
-        public ConnectionConfiguration(string endpointName)
+        ConnectionConfiguration(string endpointName)
         {
             // set default values
-            Port = 5672;
             VirtualHost = "/";
             UserName = "guest";
             Password = "guest";
             RequestedHeartbeat = 5;
             RetryDelay = TimeSpan.FromSeconds(10);
-            UseTls = false;
             CertPath = "";
             CertPassphrase = null;
 
             SetDefaultClientProperties(endpointName);
+        }
+
+        public static ConnectionConfiguration Create(string connectionString, string endpointName)
+        {
+            var builder = new DbConnectionStringBuilder { ConnectionString = connectionString };
+
+            var invalidOptionsMessage = new StringBuilder();
+
+            object value;
+
+            var useTls = false;
+            if (builder.TryGetValue("useTls", out value))
+            {
+                useTls = bool.Parse(value.ToString());
+            }
+
+            var port = useTls ? 5671 : 5672;
+            if (builder.TryGetValue("port", out value))
+            {
+                port = int.Parse(value.ToString());
+            }
+
+            var host = default(string);
+            if (builder.TryGetValue("host", out value))
+            {
+                var hostsAndPorts = value.ToString().Split(',');
+
+                if (hostsAndPorts.Length > 1)
+                {
+                    invalidOptionsMessage.AppendLine("Multiple hosts are no longer supported. If using RabbitMQ in a cluster, consider using a load balancer to represent the nodes as a single host.");
+                }
+
+                var parts = hostsAndPorts[0].Split(':');
+                host = parts[0];
+
+                if (parts.Length > 1)
+                {
+                    port = int.Parse(parts[1]);
+                }
+            }
+            else
+            {
+                invalidOptionsMessage.AppendLine("Invalid connection string. 'host' value must be supplied. e.g: \"host=myServer\"");
+            }
+
+            if (builder.ContainsKey("dequeuetimeout"))
+            {
+                invalidOptionsMessage.AppendLine("The 'DequeueTimeout' connection string option has been removed. Consult the documentation for further information.");
+            }
+
+            if (builder.ContainsKey("maxwaittimeforconfirms"))
+            {
+                invalidOptionsMessage.AppendLine("The 'MaxWaitTimeForConfirms' connection string option has been removed. Consult the documentation for further information");
+            }
+
+            if (builder.ContainsKey("prefetchcount"))
+            {
+                invalidOptionsMessage.AppendLine("The 'PrefetchCount' connection string option has been removed. Use 'EndpointConfiguration.UseTransport<RabbitMQTransport>().PrefetchCount' instead.");
+            }
+
+            if (builder.ContainsKey("usepublisherconfirms"))
+            {
+                invalidOptionsMessage.AppendLine("The 'UsePublisherConfirms' connection string option has been removed. Use 'EndpointConfiguration.UseTransport<RabbitMQTransport>().UsePublisherConfirms' instead.");
+            }
+
+            if (invalidOptionsMessage.Length > 0)
+            {
+                var message = invalidOptionsMessage.ToString().TrimEnd('\r', '\n');
+
+                Logger.Error(message);
+
+                throw new NotSupportedException(message);
+            }
+
+            var connectionConfiguration = new ConnectionConfiguration(endpointName);
+            var connectionConfigurationType = typeof(ConnectionConfiguration);
+            foreach (var key in builder.Keys.Cast<string>())
+            {
+                var property = connectionConfigurationType.GetProperty(key, BindingFlags.IgnoreCase | BindingFlags.Public | BindingFlags.Instance);
+                property?.SetValue(connectionConfiguration, TypeDescriptor.GetConverter(property.PropertyType).ConvertFrom(builder[key]));
+            }
+
+            connectionConfiguration.UseTls = useTls;
+            connectionConfiguration.Port = port;
+            connectionConfiguration.Host = host;
+
+            return connectionConfiguration;
         }
 
         void SetDefaultClientProperties(string endpointName)
