+++ src/NServiceBus.RabbitMQ/NServiceBus.RabbitMQ.csproj
@@ -75,7 +75,6 @@
   <ItemGroup>
     <Compile Include="Connection\ChannelProvider.cs" />
     <Compile Include="Configuration\ObsoleteAppSettings.cs" />
-    <Compile Include="Configuration\ConnectionStringParser.cs" />
     <Compile Include="RabbitMQTransport.cs" />
     <Compile Include="Configuration\RabbitMQTransportSettingsExtensions.cs" />
     <Compile Include="RabbitMQTransportInfrastructure.cs" />
